tests/api/requirements.txt
httpx>=0.27,<1
pytest>=8,<9
pytest-asyncio>=0.24,<1
websockets>=13,<15
cryptography>=43,<45
psycopg2-binary>=2.9,<3
